+++ asteroid/engine/schedulers.py
@@ -28,7 +28,7 @@ class BaseScheduler(object):
         for param_group in self.optimizer.param_groups:
             param_group["lr"] = lr
 
-    def step(self):
+    def step(self, metrics=None, epoch=None):
         """Update step-wise learning rate before optimizer.step."""
         self.step_num += 1
         lr = self._get_lr()
@@ -153,7 +153,7 @@ class DPTNetScheduler(BaseScheduler):
 
 
 def sinkpit_default_beta_schedule(epoch):
-    return min([1.02 ** epoch, 10])
+    return min([1.02**epoch, 10])
 
 
 class SinkPITBetaScheduler(pl.callbacks.Callback):